ADS-5332 resolve PR comments
