+++ src/Resources/app/administration/src/component/nosto-job-status-badge/nosto-job-status-badge.html.twig
@@ -1,5 +1,7 @@
 <span class="nosto-job-status-badge"
-      :class="classes">
-    <span class="nosto-job-status-dot"></span>
+      :class="classes"
+      :aria-label="label"
+      :title="label">
+    <span class="nosto-job-status-dot" aria-hidden="true"></span>
     <span v-if="showLabel">{{ label }}</span>
 </span>
