+++ src/Resources/app/administration/src/component/nosto-job-sub-jobs/component.js
@@ -4,6 +4,7 @@
 
 import template from './nosto-job-sub-jobs.html.twig';
 import fetchJobMessages from '../../util/job-messages.helper';
+import { toArray } from '../../util/collection.helper';
 import './nosto-job-sub-jobs.scss';
 
 const { Mixin } = Shopware;
@@ -16,21 +17,6 @@ const MESSAGE_COUNT_KEYS = Object.freeze({
     ERROR: 'error',
 });
 
-function toArray(collection) {
-    if (Array.isArray(collection)) {
-        return collection;
-    }
-
-    if (collection && typeof collection.forEach === 'function') {
-        const items = [];
-        collection.forEach((item) => items.push(item));
-
-        return items;
-    }
-
-    return [];
-}
-
 function normalizeMessageCountKey(type) {
     if (type === MESSAGE_COUNT_KEYS.INFO || type === MESSAGE_COUNT_KEYS.WARNING) {
         return type;
@@ -40,7 +26,7 @@ function normalizeMessageCountKey(type) {
 }
 
 function getMessageCountsFromRelations(job) {
-    if (job?.messages === undefined) {
+    if (job?.messages === undefined || job?.messages === null) {
         return null;
     }
 
@@ -91,6 +77,7 @@ export default {
             subJobs: null,
             showMessagesModal: false,
             currentJobMessages: null,
+            currentMessageRequestToken: 0,
             page: 1,
             limit: 25,
         };
@@ -214,12 +201,20 @@ export default {
             return Number(counts[MESSAGE_COUNT_KEYS.TOTAL] ?? 0);
         },
 
+        closeMessagesModal() {
+            this.currentMessageRequestToken += 1;
+            this.showMessagesModal = false;
+            this.currentJobMessages = null;
+        },
+
         showMessageModal(job) {
             const jobId = job?.id;
             if (!jobId) {
                 return;
             }
 
+            const requestToken = this.currentMessageRequestToken + 1;
+            this.currentMessageRequestToken = requestToken;
             this.currentJobMessages = [];
             this.showMessagesModal = true;
 
@@ -229,8 +224,16 @@ export default {
                 jobId,
                 expectedTotal,
             }).then((messages) => {
+                if (requestToken !== this.currentMessageRequestToken) {
+                    return;
+                }
+
                 this.currentJobMessages = messages;
             }).catch(() => {
+                if (requestToken !== this.currentMessageRequestToken) {
+                    return;
+                }
+
                 this.currentJobMessages = [];
             });
         },
