+++ src/Resources/app/administration/src/module/nosto-job/page/nosto-job-listing-index/nosto-job-listing-index.html.twig
@@ -197,7 +197,7 @@
         {% block nosto_entity_listing_job_messages_modal %}
             <sw-modal class="nosto-job-messages-modal" v-if="showMessagesModal"
                       :title="$tc('job-listing.page.listing.grid.column.messages')"
-                      @modal-close="showMessagesModal = false">
+                      @modal-close="closeMessagesModal">
                 {% block job_messages_listing %}
                     <div v-if="currentJobMessages" class="job_list_grid_job_messages-information" style="width:100%;">
                         <div class="messages-content" v-for="item in currentJobMessages">
@@ -229,7 +229,7 @@
                 {% endblock %}
                 <template #modal-footer>
                     <mt-button variant="primary"
-                               @click="showMessagesModal = false">
+                               @click="closeMessagesModal">
                         {{ $tc('job-listing.page.listing.index.close') }}
                     </mt-button >
                 </template>
