+++ src/Resources/app/administration/src/util/collection.helper.js
@@ -1,0 +1,23 @@
+/**
+ * @sw-package innovation
+ */
+
+/** @private */
+export function toArray(collection) {
+    if (Array.isArray(collection)) {
+        return collection;
+    }
+
+    if (!collection) {
+        return [];
+    }
+
+    if (typeof collection.forEach === 'function') {
+        const items = [];
+        collection.forEach((item) => items.push(item));
+
+        return items;
+    }
+
+    return [];
+}
